refactor: TableOfContents를 Loading/Loaded 합성 컴포넌트로 리팩토링
